requirements.txt
Django==2.1
djangorestframework==3.8.2
flake8==3.5.0
django-cors-headers==2.4.0